Fix what reviewing the last commit turned up, and ship 3.2.0
Ten defects, six found reviewing 6911f27 and four more found reviewing
those fixes. Each has a test that was checked against the unfixed code
first, so none of them are asserting whatever the code already did.

The two that could strand a magnet indefinitely both came from the new
metadata scheduler. Attempt budgets are per peer and per piece and
nothing expired them, so enough timeouts spent every pair and left a
download with willing peers connected and nothing scheduled - reached by
ordinary latency against the one-second default, not by malice. Budgets
are now restored when nothing is in flight and pieces are still missing.
An explicit reject survives that restoration and a timeout does not,
because re-asking a peer that already said no is how a reject storm
starts. Separately, a set that failed its info hash refused every peer
credited with an answer, and redundant requests mean most of the swarm
answers for a piece it did not supply a byte of; only the peers whose
bytes were actually stored are refused now.

Deferring pre-metadata availability had put two peer-controlled
allocations where the piece count would normally bound them. HAVE indices
cannot be range-checked before the count is known, so they accumulated
unbounded from a nine-byte message; a bitfield was kept whole up to the
2 MB message limit and pinned for the life of the connection. Both are
capped, and exceeding either discards the record rather than adopting a
partial one.

Disk speed settings stayed uint while their transfer siblings became
long, so a legal value above int.MaxValue became negative in the cast and
tripped the new negative check, failing engine startup. They are long
throughout now, which is the API change behind the version bump.

Also: HTTP tracker and web seed connects walk every resolved address of
the family instead of stopping at the first, so a tracker behind DNS
round-robin survives one host being down, bounded at three attempts since
they share one connect budget. LSD joins its multicast group on the bound
interface - binding the socket alone left membership to the routing
table, so announcements on the bound network were never seen and the join
left by another interface. And PeerDisconnectedAlert reports the client
name from the peer id rather than the endpoint it already carries.

Released as 3.2.0 rather than 4.0.0 by choice, so the changelog spells
out every changed signature and warns that a binary-only upgrade can fail
at runtime where a rebuild will not.

+++ src/PeerSharp/Interfaces/IBandwidth.cs
@@ -17,7 +17,7 @@ public interface IBandwidth
     /// </summary>
     /// <param name="torrent">The torrent.</param>
     /// <returns>A tuple containing (readLimit, writeLimit).</returns>
-    (int ReadLimit, int WriteLimit) GetTorrentDiskLimits(ITorrent torrent);
+    (long ReadLimit, long WriteLimit) GetTorrentDiskLimits(ITorrent torrent);
 
     /// <summary>
     /// Sets global download and upload speed limits.
@@ -31,7 +31,7 @@ public interface IBandwidth
     /// </summary>
     /// <param name="readLimit">Maximum disk read speed in bytes per second. 0 for unlimited.</param>
     /// <param name="writeLimit">Maximum disk write speed in bytes per second. 0 for unlimited.</param>
-    void SetGlobalDiskLimits(int readLimit, int writeLimit);
+    void SetGlobalDiskLimits(long readLimit, long writeLimit);
 
     /// <summary>
     /// Sets bandwidth limits for a specific torrent.
@@ -47,5 +47,5 @@ public interface IBandwidth
     /// <param name="torrent">The torrent to set limits for.</param>
     /// <param name="readLimit">Maximum disk read speed in bytes per second. 0 for unlimited.</param>
     /// <param name="writeLimit">Maximum disk write speed in bytes per second. 0 for unlimited.</param>
-    void SetTorrentDiskLimits(ITorrent torrent, int readLimit, int writeLimit);
+    void SetTorrentDiskLimits(ITorrent torrent, long readLimit, long writeLimit);
 }
