+++ tests/PeerSharp.Tests/Core/TorrentConfigurationTests.cs
@@ -74,8 +74,8 @@ public class TorrentConfigurationTests
     {
         private readonly Dictionary<string, BandwidthChannel> _channels = [];
         public long LastDownloadLimit { get; private set; }
-        public int LastDiskReadLimit { get; private set; }
-        public int LastDiskWriteLimit { get; private set; }
+        public long LastDiskReadLimit { get; private set; }
+        public long LastDiskWriteLimit { get; private set; }
         public long LastUploadLimit { get; private set; }
         public ITorrent? LastTorrent { get; private set; }
 
@@ -86,7 +86,7 @@ public class TorrentConfigurationTests
             LastUploadLimit = uploadLimit;
         }
 
-        public void SetTorrentDiskLimits(ITorrent torrent, int readLimit, int writeLimit)
+        public void SetTorrentDiskLimits(ITorrent torrent, long readLimit, long writeLimit)
         {
             LastTorrent = torrent;
             LastDiskReadLimit = readLimit;
@@ -107,12 +107,12 @@ public class TorrentConfigurationTests
             return channel;
         }
         public (long DownloadLimit, long UploadLimit) GetTorrentLimits(ITorrent torrent) => (0, 0);
-        public (int ReadLimit, int WriteLimit) GetTorrentDiskLimits(ITorrent torrent) => (0, 0);
+        public (long ReadLimit, long WriteLimit) GetTorrentDiskLimits(ITorrent torrent) => (0, 0);
         public Task<int> RequestBandwidthAsync(IBandwidthUser user, int amount, int priority, string[] channelNames, CancellationToken ct = default) => Task.FromResult(0);
         public void ReturnBandwidth(int amount, string[] channelNames) { }
         public void RemoveTorrentChannels(ITorrent torrent) { }
         public void SetGlobalLimits(long downloadLimit, long uploadLimit) { }
-        public void SetGlobalDiskLimits(int readLimit, int writeLimit) { }
+        public void SetGlobalDiskLimits(long readLimit, long writeLimit) { }
         public void Start() { }
         public ValueTask DisposeAsync() => ValueTask.CompletedTask;
     }
